# Improve performance / allocations of RegisterAs* methods